requirements.txt
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.0
python-dotenv==1.0.0
pypdf==3.17.4  # for PDF loading if needed
